Fix the review findings across the workspace and the vendored reader
Fifteen findings from a full-codebase review, most of them things the
suite could not reach.

The scanner: a non-ASCII byte is no longer an identifier byte on its
own. Only a letter, a digit or a combining mark is, so a Unicode
operator no longer swallows the reference after it, and the arm that
scanned non-ASCII tokens is gone — a 3-D reference whose first sheet is
non-ASCII took it and came back naming the second sheet alone, which
`audit()` could never catch because both sides call this scanner.

The store: a failed scrub used to write the deleted profiles file back,
so a corpus went on serving the cell values the caller had asked it to
forget. The manifest now goes first, as `forget` does, and a failed
`put_profiles` removes the file rather than reverting it — reverting is
a disclosure in both directions. `put` gained the rollback its siblings
have, and FORMAT_VERSION moves to 2, because the reader fixes change
what a stored graph is derived from without changing the source file.

The CLI: `--max-input-cells` defaults to no limit again. The ceiling it
carried refused the workbooks this exists for, on every verb, while
`eg serve` over the same file went on loading it.

The MCP server: an explicit null on an optional argument means absent,
which is what every `opt_*` accessor already read it as and what a great
many clients send.

The vendored reader: the `Err3d` tokens resolve through the XTI table
like their siblings (issue 9 reached only two of the four); the XLSB
3-D handlers bounds-check the index and their operand length instead of
panicking on a malformed EXTERNSHEET; an ODF error value this reader
cannot name is refused rather than reported as `#VALUE!`, which is what
both sibling readers already do; and two `as_chunks` calls go back to
`chunks_exact`, which the declared MSRV actually has.

Also: `RangeRef::rows`/`cols` widen so their product is `cell_count`
again and an un-normalised range cannot underflow; sheet access
debug-asserts on a foreign address instead of quietly reading as empty;
a defined name scoped past the last sheet is reported as workbook-scoped
with a warning rather than attributed to the first tab; and the
xls-vs-xlsx parity test is strict again, with the one real difference
between LibreOffice's two writers quarantined in a mode of its own that
fails if it is ever not needed.

Checks: the workspace suite, clippy and fmt are clean; the sweep over
the reference workbook still disagrees with nothing; its graph and
profiles are identical to the ones the previous build produced, cell for
cell; and the XLSB dump still matches the pre-change dump exactly, with
sheet-oracle agreeing everywhere it does not write columns past XFD.

+++ crates/eg-retrieve/src/expand.rs
@@ -675,7 +675,7 @@ fn cells(node: &Node) -> u64 {
     match node {
         Node::Sheet(s) => s.cells,
         Node::Region(r) => r.cell_count,
-        Node::Column(c) => u64::from(c.range.rows()),
+        Node::Column(c) => c.range.rows(),
         Node::FormulaGroup(g) => g.cell_count,
         Node::Workbook(_) | Node::DefinedName(_) | Node::ExternalWorkbook(_) => 0,
     }
